Move image controller to base project
