+++ plugins/image.js
@@ -4,10 +4,10 @@ module.exports = {
   type: 'image-controller',
   module: require('../controllers/image'),
   initialize: () => {
-    const config = require('../../collections-online/lib/config');
+    const config = require('../collections-online/lib/config');
 
     if(config.cip.client.authMechanism !== 'http-basic') {
-      const cip = require('../../services/cip');
+      const cip = require('../services/cip');
       return cip.initSession().then(() => {
         // TODO: Consider creating the structure of categories (used for the menu)
         // from another API than the CIP
